Version 1.7(Fixing a couple of errors in CreateShiftFragment),started to FullScheduleFragment

+++ app/src/main/java/com/example/onthetime/repository/EmployerRepository.kt
@@ -1,14 +1,21 @@
 package com.example.onthetime.repository
 
 import android.util.Log
+import android.widget.ImageView
+import android.widget.Toast
+import com.example.onthetime.R
 import com.example.onthetime.model.Employee
 import com.example.onthetime.model.Employer
 import com.example.onthetime.model.Location
 import com.example.onthetime.model.News
 import com.example.onthetime.model.Position
+import com.example.onthetime.model.Shift
 import com.google.firebase.firestore.FieldValue
 import com.google.firebase.firestore.FirebaseFirestore
+import com.google.firebase.firestore.SetOptions
 import com.google.firebase.storage.FirebaseStorage
+import com.google.firestore.v1.ArrayValue
+import com.squareup.picasso.Picasso
 
 class EmployerRepository {
 
@@ -79,6 +86,63 @@ class EmployerRepository {
             }
     }
 
+    fun addEmployeeToEmployeesList(
+        employerId: String,
+        employee: Employee,
+        onComplete: (Boolean) -> Unit
+    ) {
+        val employerRef = firestore.collection("employees").document(employerId)
+
+        employerRef.set(
+            mapOf("employees" to FieldValue.arrayUnion(employee))
+        ).addOnSuccessListener {
+            onComplete(true)
+        }
+            .addOnFailureListener {
+                onComplete(false)
+            }
+    }
+    fun addShiftToEmployee2(
+        employerId: String,
+        employeeId: String,
+        newShift: Shift,
+        onComplete: (Boolean) -> Unit
+    ) {
+        val employeeRef = firestore.collection("employers").document(employerId)
+            .collection("employees").document(employeeId)
+
+        employeeRef.update("shiftList", FieldValue.arrayUnion(newShift))
+            .addOnSuccessListener {
+                Log.d("Firestore", "Shift successfully added to employee's shift list.")
+                onComplete(true) // Başarı durumu
+            }
+            .addOnFailureListener { e ->
+                Log.e("FirestoreError", "Error adding shift to employee's shift list", e)
+                onComplete(false)
+            }
+    }
+
+    fun addShiftToEmployee(
+        employerId: String,
+        employeeId: String,
+        shift: Shift,
+        onComplete: (Boolean) -> Unit
+    ) {
+        val employerRef = firestore.collection("employees").document(employeeId)
+
+
+        employerRef.collection("employees").document(employeeId)
+            .update("shiftList", FieldValue.arrayUnion(shift))
+            .addOnSuccessListener {
+                onComplete(true)
+            }
+            .addOnFailureListener { e ->
+                Log.e("FirestoreError", "Error updating shift list", e)
+
+                onComplete(false)
+            }
+    }
+
     fun updateProfilePhotoPath(
         employerId: String,
         newPhotoPath: String,
@@ -91,15 +155,17 @@ class EmployerRepository {
                 if (document.exists()) {
                     val oldPhotoPath = document.getString("profilePhotoPath")
 
-                    oldPhotoPath?.let {
-                        val storageRef = FirebaseStorage.getInstance().getReferenceFromUrl(it)
-                        storageRef.delete()
-                            .addOnSuccessListener {
-                                println("Old photo deleted from storage.")
-                            }
-                            .addOnFailureListener {
-                                println("Error deleting old photo: ${it.message}")
-                            }
+                    if (oldPhotoPath != null && oldPhotoPath != "") {
+                        oldPhotoPath.let {
+                            val storageRef = FirebaseStorage.getInstance().getReferenceFromUrl(it)
+                            storageRef.delete()
+                                .addOnSuccessListener {
+                                    println("Old photo deleted from storage.")
+                                }
+                                .addOnFailureListener {
+                                    println("Error deleting old photo: ${it.message}")
+                                }
+                        }
                     }
 
                     employerRef.update("profilePhotoPath", newPhotoPath)
@@ -120,7 +186,8 @@ class EmployerRepository {
 
 
     fun getEmployerPhotoPath(employerId: String, onComplete: (String?) -> Unit) {
-        val employerRef = FirebaseFirestore.getInstance().collection("employers").document(employerId)
+        val employerRef =
+            FirebaseFirestore.getInstance().collection("employers").document(employerId)
 
         employerRef.get()
             .addOnSuccessListener { document ->
@@ -137,7 +204,8 @@ class EmployerRepository {
     }
 
     fun getEmployerName(employerId: String, onComplete: (String?) -> Unit) {
-        val employerRef = FirebaseFirestore.getInstance().collection("employers").document(employerId)
+        val employerRef =
+            FirebaseFirestore.getInstance().collection("employers").document(employerId)
 
         employerRef.get()
             .addOnSuccessListener { document ->
@@ -152,8 +220,10 @@ class EmployerRepository {
                 onComplete(null)
             }
     }
+
     fun getEmployerLastName(employerId: String, onComplete: (String?) -> Unit) {
-        val employerRef = FirebaseFirestore.getInstance().collection("employers").document(employerId)
+        val employerRef =
+            FirebaseFirestore.getInstance().collection("employers").document(employerId)
 
         employerRef.get()
             .addOnSuccessListener { document ->
@@ -168,8 +238,28 @@ class EmployerRepository {
                 onComplete(null)
             }
     }
+
+    fun getEmployerPassword(employerId: String, onComplete: (String?) -> Unit) {
+        val employerRef =
+            FirebaseFirestore.getInstance().collection("employers").document(employerId)
+
+        employerRef.get()
+            .addOnSuccessListener { document ->
+                if (document.exists()) {
+                    val password = document.getString("password")
+                    onComplete(password)
+                } else {
+                    onComplete(null)
+                }
+            }
+            .addOnFailureListener {
+                onComplete(null)
+            }
+    }
+
     fun getEmployerEmail(employerId: String, onComplete: (String?) -> Unit) {
-        val employerRef = FirebaseFirestore.getInstance().collection("employers").document(employerId)
+        val employerRef =
+            FirebaseFirestore.getInstance().collection("employers").document(employerId)
 
         employerRef.get()
             .addOnSuccessListener { document ->
@@ -184,8 +274,10 @@ class EmployerRepository {
                 onComplete(null)
             }
     }
-    fun getEmployerEmployerID(employerId: String, onComplete: (String?) -> Unit) {
-        val employerRef = FirebaseFirestore.getInstance().collection("employers").document(employerId)
+
+    fun getEmployerID(employerId: String, onComplete: (String?) -> Unit) {
+        val employerRef =
+            FirebaseFirestore.getInstance().collection("employers").document(employerId)
 
         employerRef.get()
             .addOnSuccessListener { document ->
@@ -200,8 +292,10 @@ class EmployerRepository {
                 onComplete(null)
             }
     }
+
     fun getEmployerPhoneNumber(employerId: String, onComplete: (String?) -> Unit) {
-        val employerRef = FirebaseFirestore.getInstance().collection("employers").document(employerId)
+        val employerRef =
+            FirebaseFirestore.getInstance().collection("employers").document(employerId)
 
         employerRef.get()
             .addOnSuccessListener { document ->
@@ -216,8 +310,11 @@ class EmployerRepository {
                 onComplete(null)
             }
     }
+
+
     fun getEmployerDateBirth(employerId: String, onComplete: (String?) -> Unit) {
-        val employerRef = FirebaseFirestore.getInstance().collection("employers").document(employerId)
+        val employerRef =
+            FirebaseFirestore.getInstance().collection("employers").document(employerId)
 
         employerRef.get()
             .addOnSuccessListener { document ->
@@ -232,7 +329,6 @@ class EmployerRepository {
                 onComplete(null)
             }
     }
-
 
 
     fun addProfilePhotoPath(
@@ -270,6 +366,7 @@ class EmployerRepository {
             }
         }
     }
+
 
     fun getLocationsWithSnapshot(employerId: String, callback: (List<Location>) -> Unit) {
         val employerRef = firestore.collection("employers").document(employerId)
@@ -330,8 +427,6 @@ class EmployerRepository {
     }
 
 
-
-
     fun getPositions(employerId: String, callback: (List<Position>) -> Unit) {
         val employerRef = firestore.collection("employers").document(employerId)
         employerRef.get()
@@ -351,5 +446,6 @@ class EmployerRepository {
 
             }
     }
+
 
 }
